chore: organized components
