+++ frontend/components/recent-foods/VitaminRing.jsx
