+++ frontend/tsconfig.json
@@ -29,6 +29,6 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
     "**/*.mts"
-, "components/VitaminRing.jsx", "app/vitamin-information/page.jsx", "app/vitamin-dashboard/page.jsx", "app/recent-foods/page.jsx", "components/Sidebar.jsx"  ],
+, "components/recent-foods/VitaminRing.jsx", "app/vitamin-information/page.jsx", "app/vitamin-dashboard/page.jsx", "app/recent-foods/page.jsx", "components/recent-foods/Sidebar.jsx"  ],
   "exclude": ["node_modules"]
 }
